requirements.txt
streamlit>=1.25.0
tensorflow-cpu==2.11.0
numpy
pandas
opencv-python
Pillow
matplotlib
scikit-learn
rich
